✅ 장바구니 페이지 추가, 렌더링 및 추가기능

+++ src/main/java/com/spring/mood/projectmvc/dto/responseDto/ShopItemResponseDto.java
@@ -3,6 +3,7 @@ package com.spring.mood.projectmvc.dto.responseDto;
 import com.spring.mood.projectmvc.entity.ShopItem;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +16,7 @@ public class ShopItemResponseDto {
     private Long shopItemId;
     private String shopItemName;
     private String shopItemDesc;
-    private Double shopItemPrice;
+    private BigDecimal shopItemPrice;
     private String shopItemImg;
     private LocalDateTime shopItemDate;
     private long shopItemStock;
